Kotak/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
greenlet==3.0.3
idna==3.10
playwright==1.47.0
pyee==12.0.0
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
undetected-playwright==0.3.0
urllib3==2.2.3
